requirements.txt
# Core ML dependencies
mlx>=0.5.0
mlx-whisper>=0.1.0
mlx-lm>=0.5.0

# Audio
sounddevice>=0.4.6
scipy>=1.10.0

# System integration
numpy>=1.24.0
pynput>=1.7.6
pyperclip>=1.8.2
rumps>=0.4.0

# Configuration
python-dotenv>=1.0.0
